use TestingCallback to test with a locally defined estimator class
