+++ sklearn/callback/tests/_utils.py
@@ -22,9 +22,7 @@ class TestingCallback:
         self.record = get_callback_manager().list()
 
     def setup(self, estimator, context):
-        self.record.append(
-            {"name": "setup", "estimator": estimator, "context": context}
-        )
+        self.record.append({"name": "setup", "context": context})
 
     def on_fit_task_begin(
         self,
@@ -39,7 +37,6 @@ class TestingCallback:
         self.record.append(
             {
                 "name": "on_fit_task_begin",
-                "estimator": estimator,
                 "context": context,
                 "kwargs": {
                     "X": X,
@@ -63,7 +60,6 @@ class TestingCallback:
         self.record.append(
             {
                 "name": "on_fit_task_end",
-                "estimator": estimator,
                 "context": context,
                 "kwargs": {
                     "X": X,
@@ -75,9 +71,7 @@ class TestingCallback:
         )
 
     def teardown(self, estimator, context):
-        self.record.append(
-            {"name": "teardown", "estimator": estimator, "context": context}
-        )
+        self.record.append({"name": "teardown", "context": context})
 
     def count_hooks(self, hook_name):
         return len([rec for rec in self.record if rec["name"] == hook_name])
