+++ sklearn/callback/tests/test_callback_context.py
@@ -7,6 +7,8 @@ from functools import partial
 import numpy as np
 import pytest
 
+from sklearn.base import BaseEstimator
+from sklearn.callback import CallbackSupportMixin, with_callbacks
 from sklearn.callback._callback_context import (
     CallbackContext,
     _from_reconstruction_attributes,
@@ -450,3 +452,20 @@ def test_from_reconstruction_attributes():
     assert reconstructed_est is not estimator
     assert reconstructed_est.get_params() == estimator.get_params()
     assert reconstructed_est.n_iter_ == 2
+
+
+def test_locally_defined_estimator():
+    """Test a callback with a locally defined estimator class."""
+
+    class LocallyDefinedEstimator(CallbackSupportMixin, BaseEstimator):
+        @with_callbacks
+        def fit(self, X=None, y=None):
+            callback_ctx = self._init_callback_context()
+            callback_ctx.call_on_fit_task_begin(estimator=self)
+
+            callback_ctx.call_on_fit_task_end(estimator=self)
+            return self
+
+    estimator = LocallyDefinedEstimator()
+    estimator.set_callbacks(TestingCallback())
+    estimator.fit()
